add middleware
